PreFinal

+++ server/controllers/branchController.js
@@ -2,12 +2,7 @@ var db = require('../db');
 
 //returns array of objects of branches
 exports.branch_list=(req,res)=>{
-    if(req.user.role!=="admin")
-    {
-        res.send();
-        return ;
-    }
-    const q="SELECT * FROM branch";
+    const q="SELECT branch.branch_id,branch.name,branch.location,COUNT(*) as emps FROM employee JOIN branch ON employee.branch_id=branch.branch_id GROUP BY branch_id;";
     db.query(q).then(result=>{
         result=JSON.parse(JSON.stringify(result[0]));
         res.send({
@@ -58,11 +53,9 @@ exports.post_branch=(req,res)=>{
     var id= req.body.id;
     var name= req.body.name;
     var location = req.body.location;
-    var adminName=req.body.adminName;
-    var adminPassword=req.body.adminPassword;
-    var adminEmail=req.body.adminEmail;
-    var q = 'INSERT INTO branch (branch_id,name,location,admin_name,admin_password,admin_email) VALUES (?,?,?,?,?,?)';
-    db.query(q,[id,name,location,adminName,adminPassword,adminEmail]).then(result=>{
+    var password=req.body.password;
+    var q = 'INSERT INTO branch (branch_id,name,location,password) VALUES (?,?,?,?)';
+    db.query(q,[id,name,location,password]).then(result=>{
         res.send({
             error:false
         });
@@ -86,12 +79,11 @@ exports.put_branch=(req,res)=>{
     var id= req.body.id;
     var name= req.body.name;
     var location = req.body.location;
-    var adminName=req.body.adminName;
-    var adminPassword=req.body.adminPassword;
-    var adminEmail=req.body.adminEmail;
     
-    var q = 'UPDATE branch SET name=?,location=?,admin_name=?,admin_password=?,admin_email=? WHERE branch_id=?';
-    db.query(q,[name,location,adminName,adminPassword,adminEmail,id]).then(result=>{
+    var password=req.body.password;
+    
+    var q = 'UPDATE branch SET name=?,location=?,password=? WHERE branch_id=?';
+    db.query(q,[name,location,password,id]).then(result=>{
         this.branch_id(req,res);
     }).catch(err=>{
         console.log(err);
@@ -109,11 +101,30 @@ exports.del_branch=(req,res)=>{
         return ;
     }
     var id=req.body.branchId;
-    var q = 'DELETE FROM branch WHERE branch_id=?';
-    db.query(q,[id]).then(result=>{
-        res.send({
-            error:false
-        });
+    var q1 ='SELECT * FROM employee WHERE branch_id=?';
+    db.query(q1,[id]).then(result=>{
+        result=JSON.parse(JSON.stringify(result[0]));
+        if(result){
+            res.send({
+                error:true,
+                message:"can't delete a branch containing employees"
+            });
+        }
+        else
+        {
+            var q = 'DELETE FROM branch WHERE branch_id=?';
+            db.query(q,[id]).then(result=>{
+                res.send({
+                    error:false
+                });
+            }).catch(err=>{
+                console.log(err);
+                res.send({
+                    error:true,
+                    message:'Error'
+                });
+            })
+        } 
     }).catch(err=>{
         console.log(err);
         res.send({
@@ -121,4 +132,5 @@ exports.del_branch=(req,res)=>{
             message:'Error'
         });
     })
+    
 }
